Fixing missed type conversion

+++ MAGLEV_adc.c
@@ -130,7 +130,7 @@ int median(int* arr, int len) {
 
 int16 debug = 0;
 
-int16 coilInterferenceFunc(int16 x, int dir) {
+int16 coilInterferenceFunc(double x, int dir) {
     if (!dir) {
         return (int16) (0.359427 * pow(x, 0.648402));
     }
